Remove HTML ast for HTML report generation.
