--- packages/hurl/src/report/html/ast.rs
@@ -1,52 +1,0 @@
-/*
- * Hurl (https://hurl.dev)
- * Copyright (C) 2022 Orange
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *          http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct Html {
-    pub head: Head,
-    pub body: Body,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct Head {
-    pub title: String,
-    pub stylesheet: Option<String>,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct Body {
-    pub children: Vec<Element>,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub enum Element {
-    TextElement(String),
-    NodeElement {
-        name: String,
-        attributes: Vec<Attribute>,
-        children: Vec<Element>,
-    },
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub enum Attribute {
-    Class(String),
-    //Id(String),
-    Href(String),
-    Data(String, String),
-}
